+++ core/scripts/build.sh
@@ -14,8 +14,6 @@ source "${script_dirpath}/_constants.sh"
 
 GET_AUTOUPDATING_DOCKER_IMAGES_TAG_SCRIPT_FILENAME="get-autoupdating-docker-images-tag.sh"
 
-INITIALIZER_DIRNAME="initializer"
-
 WRAPPER_GENERATOR_DIRNAME="wrapper_generator"
 WRAPPER_GENERATOR_BINARY_OUTPUT_FILENAME="wrapper-generator"
 WRAPPER_TEMPLATE_REL_FILEPATH="${WRAPPER_GENERATOR_DIRNAME}/kurtosis.template.sh"
@@ -64,10 +62,8 @@ fi
 
 # These variables are used by Goreleaser
 export API_IMAGE \
-    INITIALIZER_IMAGE \
     DOCKER_ORG \
     INTERNAL_TESTSUITE_IMAGE_SUFFIX \
-    WRAPPER_GENERATOR_BINARY_OUTPUT_FILENAME \
     JAVASCRIPT_REPL_IMAGE \
     CLI_BINARY_FILENAME
 export FIXED_DOCKER_IMAGE_TAG="${fixed_docker_image_tag}"
@@ -77,20 +73,6 @@ export AUTOUPDATING_DOCKER_IMAGE_TAG="${autoupdating_docker_image_tag}"
 cd "${root_dirpath}"
 
 go test ./...
-
-# TODO TODO TODO REMOVE ALL THIS WHEN THE CLI WILL DO EVERYTHING
-# Generate the wrapper script
-if ! goreleaser build --rm-dist --snapshot --id "${WRAPPER_SCRIPT_GENERATOR_GORELEASER_BUILD_ID}" --single-target; then
-    echo "Error: Couldn't build the wrapper script-generating binary" >&2
-    exit 1
-fi
-if ! "${root_dirpath}/${GORELEASER_OUTPUT_DIRNAME}/${WRAPPER_GENERATOR_BINARY_OUTPUT_FILENAME}" \
-        -kurtosis-core-version "${fixed_docker_image_tag}" \
-        -template "${WRAPPER_TEMPLATE_REL_FILEPATH}" \
-        -output "${WRAPPER_OUTPUT_REL_FILEPATH}"; then
-    echo "Error: Failed to generate wrapper script" >&2
-    exit 1
-fi
 
 # Build all the Docker images
 if "${should_publish_arg}"; then
@@ -103,7 +85,7 @@ if ! goreleaser release --rm-dist --skip-announce ${goreleaser_release_extra_arg
     exit 1
 fi
 
-# Build a CLI binary, compatible with the current OS & arch, so that we can run Interactive locally
+# Build a CLI binary, compatible with the current OS & arch, so that we can run interactive & testing locally
 if ! goreleaser build --rm-dist --snapshot --id "${GORELEASER_CLI_BUILD_ID}" --single-target; then
     echo "Error: Couldn't build the wrapper script-generating binary" >&2
     exit 1